responder echoes request environment
The responder echoes request environments by merging the
incoming environment with the environment created by handlers.
This results in an echo, i.e. request information is echoed back
to the client as part of the response. This issue arises for
any environment manipulating middleware, e.g. the ring cookie
middleware.